Move df concat tests to new file
